# perf: replace `detectPrng()` with `Math.random()` (up to 35x improvement)