Before:
```
cpu: Apple M2
runtime: deno 1.36.0 (aarch64-apple-darwin)

benchmark         time (avg)        iter/s             (min … max)       p75       p99      p995
------------------------------------------------------------------ -----------------------------
encodeTime       300.29 ns/iter   3,330,128.9 (296.45 ns … 330.86 ns) 300.66 ns 324.57 ns 330.86 ns
encodeRandom       5.99 µs/iter     167,070.2     (5.11 µs … 6.23 µs)    6.1 µs   6.23 µs   6.23 µs
generate           9.34 µs/iter     107,032.5      (8.89 µs … 9.8 µs)   9.61 µs    9.8 µs    9.8 µs
```

After:
```
cpu: Apple M2
runtime: deno 1.36.0 (aarch64-apple-darwin)

benchmark         time (avg)        iter/s             (min … max)       p75       p99      p995
------------------------------------------------------------------ -----------------------------
encodeTime       302.69 ns/iter   3,303,729.1 (298.42 ns … 367.81 ns) 301.97 ns 349.85 ns 367.81 ns
encodeRandom     169.59 ns/iter   5,896,426.9 (163.32 ns … 187.17 ns) 169.97 ns  179.3 ns 183.08 ns
generate         552.36 ns/iter   1,810,418.6  (544.5 ns … 590.55 ns) 554.15 ns 582.75 ns 590.55 ns
```